requirements.txt
black>=21.9b0
click>=7.0.0
Flask>=1.1.4
Flask-SQLAlchemy>=2.5.1
psycopg2>=2.9.1
pytest>=6.2.0
